+++ gui/extension/CHANGELOG.md
@@ -15,6 +15,7 @@
 - BUG#36282632 MrsBaseObjectProcedureCall instances yield a type error
 - BUG#36286261 TypeScript compiler error for accessing properties in a potential empty object
 - BUG#36347569 Links in the front page open dialog to add new REST object
+- BUG#36363355 Connection fails with scheme and non-standard port or socket in a Shell Session
 
 ## Changes in 1.14.2+8.1.1